Add Python file support in codemap and associated tests
This update includes the addition of Python file support in the 'codemap' feature. It introduces the ability to generate Python file outputs by adding it as a case in the switch statement in the 'output.go' file. It also provides new tests ensuring that this feature works correctly, through the 'output_python_test.go' file. Finally, it also adds a new 'python_output.go' file, contributing to the Python support.