docs(specs): fix typos and syntax errors in code_mode.md
This commit fixes several typos and syntax errors in the code_mode.md specification. Corrected "the execute" to "the code to execute", fixed function declaration keywords from "fun" to "func", fixed a generic type parameter from "T" to "I", and added a missing newline at the end of the file.